+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxBody.hs
@@ -45,6 +45,8 @@ module Cardano.Ledger.Conway.TxBody (
     ctbCurrentTreasuryValue,
     ctbTreasuryDonation
   ),
+  conwayTotalDepositsTxBody,
+  conwayProposalsDeposits,
 ) where
 
 import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))