+++ eras/conway/impl/CHANGELOG.md
@@ -10,7 +10,7 @@
 * Rename `PParamGroup` to `PPGroup` and `GovernanceGroup` to `GovGroup`
 * Introduce `THKD` and use it for `ConwayPParams`
 * Add `data ConwayGovEvent era` with constructor `GovNewProposals !(TxId (EraCrypto era)) !(ProposalsSnapshot era)`. #3856
-  * Add `EpochBoundaryRatifyState (RatifyState era)` inhabitant to the `ConwayEpochEvent era` data type.
+* Add `EpochBoundaryRatifyState (RatifyState era)` inhabitant to the `ConwayEpochEvent era` data type.
 
 ### `testlib`